+++ src/main/java/org/geowe/client/local/main/tool/spatial/geoprocess/dialog/GeoprocessDialog.java
@@ -32,6 +32,7 @@ import org.geowe.client.local.ImageProvider;
 import org.geowe.client.local.layermanager.LayerManagerWidget;
 import org.geowe.client.local.main.VectorLayerInfo;
 import org.geowe.client.local.main.VectorLayerProperties;
+import org.geowe.client.local.main.tool.help.HelpMessages;
 import org.geowe.client.local.main.tool.spatial.geoprocess.BufferGeoprocess;
 import org.geowe.client.local.main.tool.spatial.geoprocess.Geoprocess;
 import org.geowe.client.local.main.tool.spatial.geoprocess.IGeoprocess;
@@ -53,11 +54,13 @@ import com.sencha.gxt.cell.core.client.form.ComboBoxCell.TriggerAction;
 import com.sencha.gxt.data.shared.ListStore;
 import com.sencha.gxt.widget.core.client.Dialog;
 import com.sencha.gxt.widget.core.client.button.TextButton;
+import com.sencha.gxt.widget.core.client.button.ToolButton;
 import com.sencha.gxt.widget.core.client.container.VerticalLayoutContainer;
 import com.sencha.gxt.widget.core.client.event.SelectEvent;
 import com.sencha.gxt.widget.core.client.event.SelectEvent.SelectHandler;
 import com.sencha.gxt.widget.core.client.form.ComboBox;
 import com.sencha.gxt.widget.core.client.form.TextField;
+import com.sencha.gxt.widget.core.client.tips.ToolTipConfig;
 
 /**
  * GeoprocessDialog representa al diálogo de análisis espacial encargado de
@@ -99,14 +102,16 @@ public class GeoprocessDialog extends Dialog {
 	private LayerManagerWidget layerManagerWidget;
 	@Inject
 	private IInputGeoprocess inputGeoprocess;
+	@Inject
+	private GeoprocessHelpDialog geoprocessHelpDialog;
 
 	@Inject
 	public GeoprocessDialog(final Geoprocesses spatialOperation) {
 		super();
 		this.setHeadingText(UIMessages.INSTANCE.sodHeadingText());
 		this.setPredefinedButtons(PredefinedButton.OK, PredefinedButton.CANCEL);
-		this.setPixelSize(500, 350);
-		this.setModal(true);
+		this.setPixelSize(500, 370);
+		this.setModal(false);
 		this.setResizable(false);
 		add(createPanel(spatialOperation));
 		this.spatialOperationComboBox.setValue(null);
@@ -115,10 +120,33 @@ public class GeoprocessDialog extends Dialog {
 					@Override
 					public void onSelect(SelectEvent event) {
 						GeoprocessDialog.this.hide();
+						geoprocessHelpDialog.hide();
 					}
 				});
-	}
+		
+		final ToolButton helpToolButton = new ToolButton(ToolButton.QUESTION);
+		setHelpToolTip(helpToolButton);
+		addHelpSelectHandler(helpToolButton);
 
+		getHeader().addTool(helpToolButton);
+
+	}
+	
+	private void setHelpToolTip(final ToolButton tButton) {
+		final ToolTipConfig helpToolTip = new ToolTipConfig();
+		helpToolTip.setTitleText(HelpMessages.INSTANCE.help());
+		tButton.setToolTipConfig(helpToolTip);
+	}
+	
+	private void addHelpSelectHandler(final ToolButton tButton) {
+		tButton.addSelectHandler(new SelectHandler() {
+			@Override
+			public void onSelect(final SelectEvent event) {
+				geoprocessHelpDialog.show();
+			}
+		});
+	}
+	
 	public VectorLayer getLayer1() {
 		VectorLayer layer = null;
 		if (LAYER_COMBO_1.getValue() != null) {